pyproject.toml
[tool.poetry]
name = "tiendavirtualb"
version = "0.1.0"
description = ""
authors = ["Walter Ardila R <w.a.r15@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
